requirements.txt
# BOTA Project Requirements
# Django Framework
Django>=5.2.7,<5.3.0

# Django REST Framework and related packages
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
drf-spectacular>=0.27.0

# Filtering and CORS
django-filter>=23.5
django-cors-headers>=4.3.0

# Additional utilities (optional but recommended)
Pillow>=10.0.0  # For image handling
python-dotenv>=1.0.0  # For environment variables
polib>=1.2.0  # For compiling translation files

# PDF generation and QR codes
reportlab>=4.0.0
qrcode>=7.4.0
